% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{add_success}
\alias{add_success}
\title{Create success variable from questions}
\usage{
add_success(data, challenges)
}
\arguments{
\item{data}{Data frame returned from \code{\link[climate.campaigneRs:get_data]{climate.campaigneRs::get_data()}}.}

\item{challenges}{Data frame returned from \code{\link[climate.campaigneRs:get_challenges]{climate.campaigneRs::get_challenges()}}.}
}
\value{
List that contains the updated \code{data}, the list of challenges that
contain the question "Were you successful in completing this challenge?" and a list
of challenges that contain the question but do not occur in \code{data}.
}
\description{
Create success variable from questions
}
