% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{aggregate_challenges}
\alias{aggregate_challenges}
\title{Aggregate challenges}
\usage{
aggregate_challenges(data, challenges)
}
\arguments{
\item{data}{Data frame returned from \code{\link[climate.campaigneRs:get_data]{climate.campaigneRs::get_data()}}.}

\item{challenges}{Data frame returned from \code{\link[climate.campaigneRs:get_challenges]{climate.campaigneRs::get_challenges()}}.}
}
\value{
A data frame that contains challenges and the counts of accepted, finished,
success, fail, open, diff (difference between finished, success and fail).
Challenges are considered as identical and are therefore aggregated when all variables
except id, start and end date are identical.
}
\description{
Aggregate challenges
}
\examples{
challenges <- get_challenges()
result <- get_data(challenges)
data <- result$data
aggregate_challenges(data, challenges)
}
