% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{barchart}
\alias{barchart}
\title{Bar chart for categorical variable}
\usage{
barchart(x, useNA = FALSE, textsize = 6, labelsize = 2, title = "")
}
\arguments{
\item{x}{A categorical variable.}

\item{useNA}{A logical that indicates if NA should be included.}

\item{textsize}{Text size in pts.}

\item{labelsize}{Annotation text size in pts.}

\item{title}{Text for the title.}
}
\value{
A ggplot object.
}
\description{
Computes the proportions for a given categorical variable \code{x}.
E.g. used for socio-demographic variables such as age, gender, etc. from the
\code{data} API.
}
\examples{
\dontrun{
data <- get_data()
barchart(data$gender, useNA = FALSE)
}
}
