% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{get_challenges}
\alias{get_challenges}
\title{Access challenges API}
\usage{
get_challenges()
}
\value{
Data frame from API endpoint.
}
\description{
Access challenges API
}
\examples{
\donttest{
challenges <- get_challenges()
}
}
