% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{get_data}
\alias{get_data}
\title{Access data API and modify data}
\usage{
get_data(challenges)
}
\arguments{
\item{challenges}{Optional data frame returned from \code{\link[climate.campaigneRs:get_challenges]{climate.campaigneRs::get_challenges()}}.}
}
\value{
Data frame from API endpoint.

Modified Data frame from API endpoint.
}
\description{
If \code{challenges} is provided, it appends the variables
\code{challenges.X.success}: indicate if a challenge with ID X was finished successfully ("Yes") or not ("No").
\code{success}: Number of successful completed challenges for each user.
\code{success_bin1}: Binary variable that indicates if a user completed one or more challenges.
\code{success_bin2}: Binary variable that indicates if a user completed two or more challenges.
\code{success_bin3}: Binary variable that indicates if a user completed three or more challenges.
}
\examples{
\donttest{
data <- get_data()
}
}
