% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{get_subset}
\alias{get_subset}
\title{Get subset}
\usage{
get_subset(data, pattern = "\\\\.accepted")
}
\arguments{
\item{data}{Data frame returned from \code{\link[climate.campaigneRs:get_data]{climate.campaigneRs::get_data()}}.}

\item{pattern}{Character vector giving the patterns that can be selected.}
}
\value{
Data frame only consisting of columns with column name according to \code{pattern}.
}
\description{
Get subset
}
\examples{
\donttest{
data <- get_data()
accepted <- get_subset(data, "\\\\.accepted")
rejected <- get_subset(data, "\\\\.rejected")
finished <- get_subset(data, "\\\\.finished")
interaction <- get_subset(data, "\\\\.interaction")
co2avoided <- get_subset(data, "\\\\.CO2avoided")
success <- get_subset(data, "\\\\.success")
}
}
