% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{get_user_values}
\alias{get_user_values}
\title{Access user_values API}
\usage{
get_user_values()
}
\value{
Data frame from API endpoint.
}
\description{
Access user_values API
}
\examples{
\donttest{
user_values = get_user_values()
}
}
