% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{get_weekly_assignments}
\alias{get_weekly_assignments}
\title{Access weekly assignments API}
\usage{
get_weekly_assignments()
}
\value{
Data frame from API endpoint.
}
\description{
Access weekly assignments API
}
\examples{
\donttest{
weekly_assignments = get_weekly_assignments()
}
}
