% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_report}
\alias{render_report}
\title{Render Markdown Report}
\usage{
render_report(
  out_dir = tempdir(),
  write_excel = FALSE,
  echo = FALSE,
  shiny = FALSE,
  saveFig = TRUE
)
}
\arguments{
\item{out_dir}{A character vector of full directory name (default: \code{\link[base:tempdir]{base::tempdir()}}).}

\item{write_excel}{A logical that indicates if the API response should be written to an Excel file (default: \code{FALSE}).}

\item{echo}{A logical that indicates if code should be prevented from appearing in the final file (default: FALSE, no code).}

\item{shiny}{A logical that indicates if the report should rendered as shiny or not (default: FALSE).}

\item{saveFig}{A logical that indicates if the plots should be saved as png (default: TRUE).}
}
\value{
No return value. Called for side effects.
}
\description{
Render Markdown Report
}
\examples{
\dontrun{
render_report()
}
}
