% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{get_data}
\alias{get_data}
\title{Access data API}
\usage{
get_data(challenges)
}
\arguments{
\item{challenges}{Optional data frame returned from \code{\link[climate.campaigneRs:get_challenges]{climate.campaigneRs::get_challenges()}}.}
}
\value{
Data frame from API endpoint.

Data frame from API endpoint.
}
\description{
If \code{challenges} is provided, it appends the variables "challenges.X.success"
that indicate if a challenge with ID X was finished successfully ("Yes") or not ("No").
}
\examples{
\donttest{
data <- get_data()
}
}
